pyproject.toml
[project]
name = "projeto-inspecao"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dearpygui>=2.1.1",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
]
